refactor(readability): Phase 3 — TOML extraction and test robustness
- Extract TOML serialization into src/config/toml-serialize.ts:
  isRecord, isPrimitive, formatTomlPrimitive, formatTomlArray,
  formatTomlValue, serializeTomlSection, serializeTomlDocument.
  layeredConfig.ts imports isRecord and serializeTomlDocument from there;
  persistence.ts imports serializeTomlDocument from there.
  Removes ~95 lines of TOML formatting from layeredConfig.ts, which now
  focuses solely on config layer resolution.

- Fix fragile help-text test in handler.test.ts:
  Replace the indentation-only regex + broken comma-split logic with two
  targeted patterns — line-start (/^\s{2}\/.../) and comma-alias (/,\s*\/.../)
  — combined into a Set. This correctly picks up /quit and /settings which
  were listed after commas but silently skipped before.

No behavior changes. All 825 tests pass.
